requirements.txt
fastapi==0.136.1
groq==1.2.0
python-dotenv==1.2.2
sarvamai==0.1.28
